make two pages for grid edit/clue edit
